+++ dart/geodata/lib/src/common/links/links.dart
@@ -28,12 +28,12 @@ class Links with EquatableMixin {
   /// Creates a metadata container for links from JSON data.
   ///
   ///
-  /// Items of the [data] iterable are allowed to be `Map<String, Object?>`
+  /// Items of the [data] iterable are allowed to be `Map<String, dynamic>`
   /// (containing link attributes) or `String` (containing only a `href` part
   /// of a link).
   ///
   /// Throws `FormatException` if cannot parse data.
-  factory Links.fromData(Iterable<Object?> data) => Links(
+  factory Links.fromData(Iterable<dynamic> data) => Links(
         data.map(Link.fromData).toList(growable: false),
       );