+++ dart/geodata/lib/src/core/features/feature_items.dart
@@ -14,14 +14,14 @@ import '/src/common/meta.dart';
 @immutable
 class FeatureItems with MetaAware, EquatableMixin {
   /// Create a feature items instance with [collection] and optional [meta].
-  const FeatureItems(this.collection, {Map<String, Object?>? meta})
+  const FeatureItems(this.collection, {Map<String, dynamic>? meta})
       : meta = meta ?? const {};
 
   /// The wrapped feature collection.
   final FeatureCollection<Feature> collection;
 
   @override
-  final Map<String, Object?> meta;
+  final Map<String, dynamic> meta;
 
   @override
   List<Object?> get props => [collection, meta];