+++ dart/geodata/lib/src/ogcapi_features/model/ogc_feature_items.dart
@@ -35,7 +35,7 @@ class OGCFeatureItems extends FeatureItems with LinksAware {
   @override
   Links get links {
     final data = meta['links'];
-    if (data is Iterable<Object?>) {
+    if (data is Iterable<dynamic>) {
       try {
         return Links.fromData(data);
       } on FormatException {